pyproject.toml
[project]
name = "vbrownbag-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.84.0",
    "fastapi>=0.134.0",
    "jinja2>=3.1.6",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.41.0",
]
